zerotier improvements + introduce Devspace environment
Zerotier improvements
=====================

The current zerotier implementation creates just one IP per cluster.

We are proposing a model where every tezos node gets its own ip.

Advantages:

1. it removes any routing and forwarding consideration - every node is directly connected to a virtual layer 2 switch
1. it makes no assumption on whether a tun interface can be shared between pods (which is not true on i.e. eks)

Implementation: when a zerotier network is passed, we assume zerotier is in use. We append 2 containers to the nodes: one init container which gets an ip from zerotier, and one container that runs the zerotier daemon and runs alongside the tezos node. The tezos node is configured to "think" that its ip is the zerotier ip.

We give predictable names to zerotier interfaces <CHAIN_NAME>_bootstrap. When a cluster is invited, it queries all interfaces, and uses the ones named as such as bootstrap nodes.

This removes the requirement to query the ip of the zerotier interface and pass it as an argument to the invite command.

introduce devspace environment
==============================

Zerotier container used to be in docker hub. It was based on https://github.com/leunamnauj/kubernetes-zerotier-bridge

We moved this source code into the current repo and modified it. Building the container is a prerequisite of running a private chain with zerotier enabled.

We introduce devspace to facilitate builing containers in a minikube environment. Devspace is a tool to facilitate development in a kubernetes cluster. It monitors files, and when they change, rebuilds containers and restarts pods.

README changes
==============

README gives instructions for the single-cluster setup

HOWTO is renamed to MULTICLUSTER.md and is the guide for multi-cluster including zerotier

Note
====

The invite process is made more complicated by this change, because now it requires the complete mkchain + devspace env.

This satisfies the immediate need of multiple developers collaborating on chain developent.

Future work will simplify this process. This may consist of releasing custom containers to a public registry as part of a release cycle.