Release pipeline
To release mkchain, create a Github release with PEP440 compliant release tag (for example `1.0b1`, `1.0`)

The following will happen:

* PyPI release of the package
* all containers in `docker` folder will be pushed to docker hub

Example trigger of the container build pipeline:

https://github.com/tqtezos/tezos-k8s/actions/runs/363520964

The `image` field for custom containers in the k8s yaml output will point to the corresponding version on docker hub. There is no need for end users to build containers: sending a k8s yaml as private chain invite will just work. This applies to pip installs from pypi, pip installs of a tagged github url, and manual cloning of a tagged version followed by `pip install -e .`.

We are not using docker hub build pipelines, instead using github actions. This is more sustainable as it reduces the amount of work needed to migrate to another registry.

This was done with a python utility named [versioneer](https://github.com/python-versioneer/python-versioneer). versioneer.py and tqchain/_version.py are files from this utility, and must be committed in our source tree for this to work.

Also added `mkchain --version`

When this PR is approved and merged, I will release 1.0b1