Move requirements.txt out of rpc-auth/server
This is really because of devspace sync. It takes a directory and you can't specify a specific file to include or exclude. So it uploads everything in the dir to the container. Therefore there are sync permission errors when it tries to upload files to the container that don't have permissions.