edo fix + switch kubernetes ingress to beta + switch tezos to v8
* Edo fix:

The activate job used to do 2 things: activate protocol and bake once.
At some point, it was necessary to bake once manually to kick the chain
off, but this requirement has now been lifted. After activation, new
blocks will be producing automatically.

That works for delphi but it went unnoticed. However for edo the protocol
is now more picky in the time distance between blocks so bake_once fails.

I am removing it. It works on delphi and edo still, but carthage may
be broken.

In the process, I am creating a specific container for activation, because
a bash script is easier to understand and debug than a command in a
kubectl manifest.

Per our convention, this new container goes at the top level, although
this is maybe worth discussing because it is so simple it will never
become its own project.

Also updated tezos to v8. Delphi and edo work.

* Switch kubernetes efgress to beta:

This is necessary to make ingress-nginx work on EKS.

The syntax is different (serviceName/servicePort). I tested on minikube
as well.

Also replace readWriteMany (unsupported on cloud) with readWriteOnce