support for multiple bakers
Previously we used to create 4 accounts always. We simplify by
creating just one account per baker. As a consequence, all blocks
are baked on priority zero from the beginning.

We add a new parameter to mkchain `--number-of-bakers`, defaults to
1. If we pass a higher number n, it will create n bakers and n nodes
baking for each baker.

Bakers and nodes can be mapped using values.yaml: specify which node
bakes for which address. This requires jq which requires a custom
baking container. This also allowed me to remove helm hacks.

We make a few assumptions for now:
* every baking node is also a bootstrap node
* the first baker's address is also the chain activator

I tested in non-zerotier mode and zerotier mode with one create
namespace and one invite namespace in the same minikube cluster; it
works, but required me to remove the hard-coded nodePort for p2p.