wait for bootstrap node to be up before launching other nodes
We replace the bootstrap service with a headless service and have
all nodes connect directly to the pod ips.

This exposes a bug where nodes are stuck with zero connections if
they try to connect too early. This was previously masked by kubeproxy.

We fix that by adding an init container which waits for the first
bootstrap node to accept connections.

The bootstrap attribute is now declarative: we no longer consider
every baking node as bootstrap. The bootstrap: true flag must be passed
in values.yaml. mkchain adds it to the first two baking nodes.

We add headless services in order to give the pods a meaningful FQDN.

Rename "bootstrap" statefulset to "baker" to reflect this new model.

Other changes:

* In values.yaml, baker nodes and regular nodes are now listed
separately, which simplifies the code in other places.