Only import the required keys in init containers
We do import all of the keys into the activate-job, because it needs
to know about all of the accounts in order to set them up with a init
balance.

We also stop generating /etc/tezos/parameters.json except on our
activate-job because that's the only place where it is necessary.