Remove readiness probes
Readiness probes have the unfortunate side-effect of causing the
DNS resolution of our hostnames to fail whenever kubernetes thinks
that the node is not ready.  This exposes some underlying bugs in
tezos-node and causes many nodes to simply stop trying to find
peers.  Waiting to start helps the issue, but we find that minikubes
sometimes decides that nodes are unavailable randomly even after it
considered them to be ready.  This makes our name service complete
non-deterministic.