Update and simplify pulumi image locations, generation code
We update the images names to those used at the head of the current
master branch.  We also start to use a small loop to generate the
locations of the docker images, this reduces repetition.

pulumi should also use the tezos namespace.

Choose the size of the cluster based on the number of tezos nodes.

pulumi/index.ts: make rpc-auth and nginx conditional.

Add a README.md to the pulumi directory.

Choose the CHAIN_NAME from the env var rather than hard-coded.

Put out chain name into the objects that we create.

General cleanup.