baker/endorser containers must wait for node to start
This solves a race where the baker/endorser may exit if the node has not
yet started when it starts.  Of course, k8s will restart the container,
however, this gives us RESTARTS on the console which we are hoping will
exclusively represent issues with the underlying code rather than our
test framework.