Decouple number-of-nodes from number-of-bakers (#100)
* Decouple number-of-nodes from number-of-bakers

We now consider --number-of-nodes to be the nodes which do not
include a baker/endorser pair.  It's more natural to consider
that:

    $ mkchain --number-of-bakers 40 --number-of-nodes 5 my_chain

will bring up 40 baker nodes and 5 plain nodes.

This is helpful with pulumi because it has been waiting for zero nodes
to come up and timing out.  With the current code:

    $ mkchain --number-of-bakers 40 my_chain

will bring up 40 bakers and a single node.

* Allow either zero bakers or nodes but not both.

* Update docs

* mkchain/README.md: Improve readability

* Doc updates

Co-authored-by: Aryeh Harris <harryttd@gmail.com>