Fill in defaults aggressively in the containers
We now import the keys in the config-generator container because
we use Python to construct the contents of /var/tezos/client by
doing a little crypto.  We generate deterministic keys based on
the account names for keys that were not provided and intuit accounts
using default names where they are not provided.

Keys are used if they are provided.

We eliminate the key-importer container because we no longer use it.