add Nomadic's tezos indexer
It works with `mkchain --chain-type public --network edo2net`

Issues:

1. it does not work with a private chain because the indexer does
not have this feature. Specifically, blocks 0 and block 1 are hardcoded. A workaround is to
write block 0 and 1 as part of the intial db migration, but the
hash of block0 comes from values.yaml, and the hash from block 1 is
different every time. Tezos-indexer team is working on private chain
support.
1. the schema is hardcoded in the commit because it's simpler. The
indexer container should probably deploy the schema at first boot.