Various changes to get zerotier working (not the invite)
The most interesting one here is that for some reason, zerotier
races with tezos-node.  If the latter wins, then it fails to bind
to the p2p listen-address because zerotier hasn't yet setup the VPN.
We solve this by restarting tezos-node a number of times on failure.

We also sleep 3600 on tezos-node exit to give us a chance to get a
shell on the node.  We may want to consider this, though.