pulumi/index.ts: Fix calculation of necessary nodes in AWS.
When we updated our representation of the network, we forgot to
update the calculation in pulumi/index.ts of how many pods we are
creating.  The data structure changed from a list to an object and
hence the length function returned zero leading to us calculating
that we need the minimum number of machines regardless of the number
of bakers and nodes we are provisioning.