remove RPC auth from dependencies of tezos-k8s
Pulumi won't deploy the tezos chart from submodule when the dependency exists
(even if its condition is set to false).

To enable rpc-auth, it is now necessary to explicitly install the chart.

Since the `rpc_auth` condition in tezos helm chart was only used for the
conditional rpc-auth dependency, we remove it everywhere.