Use the provided image for all tezos binaries (#177)
* Use the provided image for all tezos binaries

To do this, we switch from making containers based on
tezos/tezos:v8-release to making containers that either:

	1. use the provided tezos image and execute
	   a script which is passed in via the
	   statefulset configuration, or

	2. are not based on the tezos image and do not
	   contain tezos binaries.  These containers
	   prepare configuration, files, etc., for the
	   later containers.

* Rename config-generator to "utils" and fold in wait-for-bootstrap

As a first step on consolidating to fewer docker images, we fold
wait-for-bootstrap into config-generator calling the result "utils".

* Make utils/entrypoint.sh into a spawner of the other scripts.

And fix a few things along the way.  I think that I'll squash this
into the previous commits before merging to the master branch.

* Forgot to git add utils/config-generator.sh.

* Rename file to correct place.

* Make tezos-node also use an external script.

This lets us put some debugging in there.

* utils/Dockerfile: pin to python:3.8-alpine and streamline.

* Various changes to get zerotier working (not the invite)

The most interesting one here is that for some reason, zerotier
races with tezos-node.  If the latter wins, then it fails to bind
to the p2p listen-address because zerotier hasn't yet setup the VPN.
We solve this by restarting tezos-node a number of times on failure.

We also sleep 3600 on tezos-node exit to give us a chance to get a
shell on the node.  We may want to consider this, though.

* Fix zerotier invites...

The issue was that the CHAIN_NAME env var wasn't percolating to
the inviter and so the host was labeled _bootstrap.  The invitee
then looked for ${CHAIN_NAME}_bootstrap, which obviously wasn't
there...

* Eliminate spurious error.

* Use Python 3.9 rather than 3.8.

* Remove confusing comment.