do not start a tezos-node by default
for teztnets, I only deploy a baking node. But the regular node is always
there and it is causing issues with deployments.

I suppose helm does a deep merge of the values and passing `regular: {}`
does not help.

This is a draft for now; I am using this submodule to deploy teztnets.
This may require more digging into helm behavior to see if it is
avoidable. By default, helm deploys a mainnet node when no values.yaml
is passed. This is desirable and should be kept.

Otherwise this will require a doc fix and deploying a mainnet node will
be slightly more complicated as it requires a non-empty values.yaml