Add a useful tool that compares the charts generated by...
...two different git commits.  This allows you to do a bit of
development and compare what your charts look like compared to
the master branch of the tree:

	:; helm-compare-charts my_chain_values.yaml

You do need to specify a values file.

Your tree is required to be clean, because we checkout the master
branch.  We could make a mode where it stashes your changes and
compares before and after what you're working on.