genesis contracts and commitment URL support + remove hard-coded commitments (#196)
Commitments are precursor to faucet accounts. we had previously committed them in here for the sake of getting florencenet out, but they really belong in teztnets (and in fact the faucet itself belongs there too).

It is not possible to pass such a large object with helm values, so we implement a mechanism to download them from a random http location. On teztnet's side, I am putting this in source code and it gets deployed in a s3 bucket whose url is then passed to helm

A mechanism to inject contracts at genesis is also included. It works very similarly to the commitments. example

Also changed logging behavior a bit: when passing commitments, parameters.json is so large than debugging is impossible. I am no longer printing it to the logs, and I also limit the output of activation to 200 lines.

Granadanet will be deployed this way so it's a good idea to merge soon.

(I will squash and merge since commit history is messy)

* genesis contracts support

* make commitments a url too (move actual file to teztnets)

* remove COPY

* remove old way of injecting commitments

* wrong log output

* allow passing of arbitrary object to node config

allow arbitrary config from values.yaml to be set in the node. This
allow for setting private mode from values.yaml and this setting is
transparently passed to the node config.

* remove change that should not be in this PR

* add requests dep