remote signer statefulset
This creates a new values.yaml object `signers` deploying `tezos-signer` daemons in pods in dedicated k8s statefulset.

Signers are associated with accounts with the syntax: "sign_for_account": list of accounts.

When a signer is found with relevant "sign_for_account", the baker will automatically configure itself to use it.

Example values.yaml config:

```
nodes:
  baking:
    tezos-baking-node-0:
      bake_using_account: baker0
      is_bootstrap_node: true
    tezos-baking-node-1:
      bake_using_account: baker1
      is_bootstrap_node: true
  regular: null
signers:
  tezos-signer-0:
    sign_for_accounts:
    - baker0
    - baker1
```

In this PR, we also limit the private key exposure to the containers where they are needed. Private keys only go to the baker if needed, or to the signer if needed, based on values.yaml config.

mkchain will deploy one signer by default, and every baker will connect to it. This is made standard as this is good practice.

We also replace `MY_NODE_NAME` with `MY_POD_NAME` since the signer also populates this env var in order to help config generator to generate the right config.

Previously, config-generator.py relied on `MY_NODE_NAME` being unset to detect whether it was running inside the activation pod. We have now made it explicit by setting `MY_POD_NAME` to `activating` in the activation pod.

This PR also fixes some linting issues that are in master now.