optionally use tezedge
This is alpha code to optionally use tezedge.  If you specify "tezedge"
in the "runs" list of node definition, then it will run tezedge rather
than a tezos-node.

We have an existing bug which is that for some reason, we are not getting
DNS resolution correctly in the tezedge container supplied to us, so it
doesn't quite work.  If you modify the code and plunk in IP addresses
in the appropriate part, though, it does work, so this is close.