utils/config-generator.py: use remote signing URL if MY_POD_TYPE...
...is not "activating" or "signing" rather than if it is "baking".
I refactor the code a little bit along the way to make the intentions
more clear.

This change also fixes a bug that I would have introduced in a later
commit had I not fixed it, namely I am going to eliminate the difference
between a "baking" and "regular" pod and they will be different
configurations of a new "node" pod.