Dynamic faucet support
Previously, we supported injection of commitments from a json url.

This json url was exposing a s3 bucket with the same faucet that was
used for all tezos testnets historically.

The current commit allows for one faucet per testnet.

A faucet consists of 2 files:

* a public commitments file listing the precursor tezos addresses that
  turn into tezos addresses when the account is claimed (using the same
  method than used on mainnet to claim ICO contributions)
* a private file listing all the secret keys for all the faucet
  accounts.

The public commitments file is added to the parameters.json file used
for chain activation, the private file is not needed here so is just
disregarded.

We took the script generating these 2 files from the tezos repo, made it
deterministic, and put it here.

So the following activation section in values.yaml will always generate
the same faucet addresses:

```
activation:
  deterministic_faucet:
    seed: hello!
    number_of_accounts: 99
```

This PR is a subset of the work in teztnets where the seed is generated
based on a secret string concatenated with the full name of the testnet.
It is then injected to the helm values for the tezos-k8s instance.
Teztnets also deploys the faucet website itself, which is running the
same script and is passed the same seed.