fix #227 handle the new ACL setting (#228)
To run octez master branch, it's necessary to put the following setting
in values.yaml

```
open_acls: true
```

This will configure the baking nodes as follows:

```
  "rpc": {
    "listen-addrs": [
      "172.17.0.7:8732",
      "127.0.0.1:8732"
    ],
    "acl": [
      {
        "address": "172.17.0.7",
        "blacklist": []
      }
    ]
```

The above configuration is needed for chain activation, since the
activation operation uses RPCs which are blacklisted by default.

The ACL setting needs to be set with the listen addr, so it's not
possible to pass it as a node config with values.yaml. It has to be
generated on the fly. We pass the flag to config-generator.py which adds
the setting.

When set to `true`, this toggle will break octez v9/10 since it will not
recognize an `acl` object in `config/rpc`.

The toggle will go away at some point (when v9/v10 is no longer used)