wait-for-bootstrap.sh: wait for name service only
We only wait for name service to become available because this is all
that we actually need.  Only waiting for name service will avoid potential
deadlocks where two nodes are waiting for each other.

We also simplify the script by using the list of peers in
/etc/tezos/config.json rather than parsing $NODES with jq.  It is much
easier to understand the jq and, more importantly, we don't want to
duplicate the logic of determining the node's list of bootstrap peers,
because as we define network topologies this logic is likely to become
substantially more complex.  We did have to move the wait-for-bootstrap.sh
container to after config-generator.sh to accomodate this.

Fixes #216.