Readiness probe for tezos-node (#283)
We already had a readiness probe for tzkt, this PR adds it for the node
itself.

This is done by a python sidecar. When it receives a get request on the
/is_synced endpoint, it will query the node head block header and verify
timestamp is less than 10 minutes old.

We are using a http probe served by a flask/werkzeug server. I had to
add flask to the utils container.

When the chain is not bootstrapped, it is considered healthy so the
activation call can reach it.

This is important for a RPC server with several endpoints as they may
not all be synced. For example, when you just created the statefulset,
they sync in sequential order.

More importantly, in combination with #282 health check becomes more
important. As the chain syncs from snapshot or tarball each time it is
launched, it is important to not direct RPC traffic to nodes that are
still syncing.