switch genesis proto to mainnet-genesis (#295)
It looks like since the beginning of tezos-k8s we had been using the
`PtYUEN` protocol which is carthage-bootstrap.

Nomadic is going to remove this protocol from tezos source so they asked
that we switch to tezos mainnet activation. Doing this here for our
mkchain-generated values.yaml files and the commented out example in the
default one.