deprecate endorser binary (#316)
* deprecate endorser binary

Mondaynet starts on Hangzhou then upgrades to alpha. Hangzhou has an
endorser but alpha does not. Currently there is no way to handle this.
This PR fixes it.

The endorser days are numbered, so instead of giving people to start or
not start an endorser based on the protocol, I just hardcoded this
simple logic:

  Start endorser if granada or hangzhou, else not.

* fix mkchain