Add a chart test framework.
Essentially, the idea is that we have a directory which contains
a set of values files: test/charts/*.in.  Corresponding to each chart
is the fully expanded chart that they generate.  This is stored in
the same directory with the .in replaced with a .expect.  They are
both checked into git.

We then provide a script called test-charts which uses "helm template"
to expand our charts using each values file and makes sure that it
matches the expected chart.  The script will error out if there are
any differences.  It will also emit the diffs.

If the diffs are acceptable, just:

	:; ./bin/test-charts -r <values_file_name>

will put the new expected chart output in place where you can add
it to your commit.