add ability to pass node identity as values
with the advent of tarballs, spinning up tezos nodes with local storage
is very fast. But each time it starts, it generates a new identity, which
requires solving a proof-of-work puzzle which takes few seconds to 1-2
minutes.

You can now generate these identities out-of-band, and record them in
values.yaml at node level, and save this time at boot of your local
storage-backed node.

This is not supported at node_class or global level since each identity
must be unique.

A controversial choice I made is to add these inside the tezos-config
configMap instead of a secret. Thinking about it, these identity keys
are not that secret, even though they use public key cryptography and
have a "secret" key. The impact of a leak is that you can spoof a peer
and potentially knock it off the network. It is not bad enough that we
would want to create a k8s secret. Due to helm limitations, we would have to
make a new top-level object in values.yaml especially for secrets, which
must match the structure of `nodes`. This would not be great for
usability.

That being said, I add a `set +x` around the bottom of
config-generator.sh so we don't unnecessarily print this key in the
logs.