make tezos-client command work without passing client dir as param (#369)
This is very neat because I can open a shell into the node or baker and
type `tezos-client list known addresses` and it just works.

Q: why a symink instead of moving the actual dir to
/home/tezos/.tezos-client ?

A: the potential for breakage is high if we move the dir. This makes the
startup script a bit longer but there is virtually no risk of breakage.