make readiness probe optional
For Teztnets, it's useful to disable this probe: when I start a new
testnet, sometimes I am waiting for other participants to join for the
chain to kick off.

When recreating dailynet or mondaynet, I need to wait for other peers to
connect to me to download the most recent blocks: with a probe, that's
not possible.

Must be explicitly set to false, otherwise by default the probe will be
present. It can only be set at node type level. Since it's processed by
helm and not config-generator, I didn't see a way to make it
configurable at global or instance level.