make hard-coded identity work when snapshot is being imported
Snapshot import wipes the data dir. But we may write an identity.json
file in it. This PR ensures that we are not deleting it after writing
it!

We also write it at every boot (because what if the user wants to change
it?)

We add the node id in the debug log for config-generator.