handle new mandatory liquidity baking vote in alpha (#392)
* handle new mandatory liquidity baking vote in alpha

The new protocol now turns the liquidity baking escape vote into an escape
toggle and make it mandatory.

This broke tezos-k8s, and consequently dailynet and mondaynet.

According to tzip, third-party baking software is encouraged to not
provide any default to the user, so they are forced to take a stand.

There are two ways to vote, either pass --liquidity-baking-toggle-vote
on/off/pass or pass a --votefile with your vote in json format.

Even when a votefile is passed, --liquidity-baking-toggle-vote is
mandatory, which I think is bad UX. I [opened a
ticket](https://gitlab.com/tezos/tezos/-/issues/2641)

My plan, when this ticket is fixed, is to allow tezos-k8s users to
simply pass structured vote data as a value to "runs" "baker".

For now I am leaving this PR as draft.

For example:

```
nodes:
  rolling-node:
    runs:
    - octez_node
    - baker:
        vote:
          liquidity_baking_toggle_vote: pass
```

This structured data is then passed to the baker daemon as a votefile.

This satisfies the requirement that tezos-k8s does not default to "pass"
and forces the user to take a stand.

This will allow us not to write special case satements in the baker
entrypoint script to pass this or that flag based on which protocol we
are running...

For now this draft PR is just a fix to make dailynet and mondaynet
deploy. I'll deploy it to teztnets as a submodule.

* do not pass toggle vote for ithaca