aggressive refactoring of signer
This is a continuation of @elric1's aggressive refactoring of the node
templating. We are defining signers using _containers.tpl in a generic
way. This results in less code and easier to understand. Plus, we get
the ability to set env variables "for free".

The env var change was my original intent: I need to parametrize magic
bytes differently based on the network type. Instead of making a
custom parameter, I allow a generic $SIGNER_EXTRA_ARGS parameter which I
can use to set magic bytes or anything else.

This results in a change in values.yaml (and will require a major
version bump). We are now defining signers exactly like nodes, as a map
of statefulsets. Each statefulset contains an `instances` list where we
define values custom to each pod:

```
signers:
  tezos-signer:
    env:
      all:
        SIGNER_EXTRA_ARGS: --magic-byte 0x11,0x12,0x13
    instances:
    - sign_for_accounts:
      - baker0
```

While previously it was:

```
signers:
  tezos-signer-0:
    sign_for_accounts:
    - baker0
```

This allows signers to be configured exactly like nodes, which is more
intuitive for the tezos-k8s user.