Set per-block-votes from values.yaml
This gives a tezos-k8s baker the ability to set per-block votes in the
`protocols` section of values.yaml

It is also made mandatory: when this is unset, the baker will give a
useful error message.

This fits the tzip recommendation of not giving a default choice to the
user.

I verified with [MR 5390](https://gitlab.com/tezos/tezos/-/merge_requests/5390) that it
works properly: when set, the baker starts. When unset, it gives the
user suggestions on how to set it.

For now, voting is global. Custom votes per baker will be supported
later, by allowing an override of `protocols` at statefulset instance
level.

I am leaving as draft. We need to wait for 13.1 to release this.

Note: when we move mkchain to support jakarta by default, we should set
a pass vote. For private chains, it's fine to set a default.