Remove documentation in README, redirect to tezos-k8s.xyz
I have done a quick review, everything from the README was in the doc
except the section about indexers and the one about using a custom tezos
image. I have added them.

tezos-k8s.xyz is officially in MVP status, therefore it is good to send
people to it now, and improve it as we go.