Generate a genesis block when omitted from config (#470)
* Generate a genesis block when ommitted from config

The chain id is deterministically generated from the genesis block hash,
and it is important that every new public chain has a different one.
On teztnets, we accidentally gave the same block hash to half-baked
kathmandunet and jakartanet.

To prevent this in the future, we add a tezos-k8s feature that generates
a genesis block hash when absent. It is derived from chain_name.

@elric1 had already written some code for that but it was gated by a
flag `should_generate_unsafe_deterministic_data`. The purpose of that
flag is to autogenerate deterministic well-known accounts for private
simulations. It also generates the genesis block hash, but that's not
unsafe per se. So I moved this code out of the flag. When the chain_name
is provided, it's using it as the seed, otherwise it uses "foo" as
before.

Mkchain now outputs a values.yaml without genesis block. It will be
generated with the method introduced here.

* address reviews:

1. handle public chains
2. remove flextesa function totally
3. get(something, None) => get(something)

* remove unused import from keys

* per review, fail if chain_name missing