podman compat in development mode
Podman does not accept docker image tags in the format image:tag as
local, it requires `localhost/` when the image was built locally.

This small change saves me time because a `mkchain` default yaml works
on my machine again.

I believe this will not break the development workflow for docker users,
please confirm.