export all artifacts in jekyll metadata
This allows for the creation of a file explorer.

A basic file explorer can be deployed by merging the PR:

https://github.com/oxheadalpha/xtz-shots-website/pull/14

Before this goes to mainnet, this file should probably be renamed from
`getLtestSnapshotMetadata.py` to `generateJekyllData.py` since it's no
longer only about latest artifacts.

When we have v13 and v14 namespaces, the v13 and v14 artifacts will
seamlessly be mixed together in the same file explorer page. This is as
desired.

After we deploy the file explorer and v13/v14 artifacts, let's send this
tweet:

> Tezos Kathmandu protocol will activate on <date>. You must upgrade your
node to Octez v14.
> Upgrading your node requires a storage migration step. To make
migration easier for you, snapshots and tarballs for both v13 and v14
are now available on XTZ-Shots.io
> You can easily browse all available artifacts with our new file
explorer:

Attach a small video of the front page and file explorer navigation flow
to the tweet.